rename links and renderer, add rounded corners
