+++ app/lib/rolemodeler/draw/ROMRendererUtil.js
