+++ app/lib/rolemodeler/features/ordering/ODOrderingProvider.js
@@ -25,7 +25,7 @@ export default function ODOrderingProvider(eventBus, canvas, translate) {
   var orders = [
     { type: 'rom:Role', order: { level: 5 } },
     {
-      type: 'rom:Inheritance',
+      type: 'rom:Link',
       order: {
         level: 3,
       }
