+++ app/lib/rolemodeler/features/rules/ODRules.js
@@ -206,7 +206,7 @@ function canConnect(source, target) {
     return null;
   }
   if (canConnectLink(source, target)) {
-    return { type: 'rom:Inheritance' };
+    return { type: 'rom:Link' };
   }
   return false;
 }
